Let exported mal function return int instead of void.

+++ sql/test/configurable_rss/Tests/rss.py
@@ -9,7 +9,7 @@ class RSSTestConfig:
     sql_template = \
 """
 create function test_rss(bound bigint)
-returns bigint external name test_config_rss.run_test_config_rss;
+returns int external name test_config_rss.run_test_config_rss;
 
 select test_rss({gdk_mem_maxsize});
